services/maps-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
prometheus-client
psycopg2-binary
python-jose[cryptography]
